simple-todo-app/app/requirements.txt
asgiref==3.7.2
Django==3.1.3
djangorestframework==3.12.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
